remove trivial numeric casts

+++ src/main.rs
@@ -61,7 +61,7 @@ fn main() {
 
         // Write the image to output.png.
         match image::save_buffer("output.png", &img,
-                                 width as u32, height as u32, image::RGB(8)) {
+                                 width, height, image::RGB(8)) {
             Ok(_) => println!("wrote image to output.png"),
             Err(reason) => println!("failed to write output png: {}", reason)
         }
